+++ build_tools/python/e2e_test_framework/serialization.py
